Fix positioning bug in iOS Safari/WKWebView
I believe this was caused by a WebKit bug[1] that is triggered when we
set the document body to position fixed and then try to measure element
position on it. Since the element position doesn't actually change
before and after fixing the document, I believe it's safe for us to grab
it before we adjust the document, so that we don't run into that bug.

[1] https://bugs.webkit.org/show_bug.cgi?id=176053