Add support for `popover` attribute
This is better than the `position: fixed` hacks we were doing before
because it ensures that the open menu is in the top layer and cannot be
obscured by other content on the page (such as absolutely-positioned
elements with high Z-index values).

We set up the popover to be manually controlled, so we still retain
control over the styling, positioning, and how/when it opens and closes.